requirements.txt
pandas>=2.0.0
openpyxl>=3.1.2
xlsxwriter>=3.1.0
pyyaml>=6.0.1
httpx>=0.26.0
tenacity>=8.2.2
pydantic>=2.5.0
python-dotenv>=1.0.0